fix(engine): make logs tolerate a replica-resolution failure per service
The target-resolution loop in `logs_with_options` called
`live_replica_names(n, s).await?`, so a transient libpod error resolving
one selected service's replicas aborted the whole command before a single
log line printed - even though the same function already documents (and
the per-container streaming loop below it already implements) tolerating
a missing/not-yet-created container so the services that do exist still
stream. I now match on the per-service result, warn and skip a service
whose resolution errors, and continue with the rest, so `logs a b c`
still streams a's and c's logs when b's resolution fails.

Left a comment pointing at scale.rs's `list_project_containers_by_service`
bulk primitive as a future optimization over the current one-round-trip-
per-service loop - out of scope here, this change is resilience only.

Added a test with two selected services where the fake libpod 500s on
one's container-list lookup and returns a live container for the other;
confirmed it fails against the pre-fix `.await?` (the whole command
errors and the healthy service is never reached) and passes after the
fix. Moving mod.rs's inline test module out to a sibling `tests.rs` to
add this test kept it under the 500-line hard limit, the same split
already used for ps.rs and the lifecycle module.
